+++ CHANGELOG.md
@@ -4,16 +4,21 @@
 - **Fix** : fix customer mails not received
 
 ## 2.27.5
+
+- **Add** : Added new iDeal logo
 - **Fix** : add reference_to_pay column if missing
 
 ## 2.27.4
+
 - **Fix** : add upgrade script
 
 ## 2.27.3
+
 - **Fix** : Fixed error when opening order details
 - **Fix** : Fixed capture operation displayed twice
 
 ## 2.27.2
+
 - **Fix** : Fixed missing refund button when language changes
 - **Fix** : Fixed Apple Pay card brand detection using payment_product property