+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## UNRELEASE
 - **Fix** : fix customer mails not received
+- **Add** : Fixed product customization incrementation error
 
 ## 2.27.5